fixed Randomwalker, bandmatrix errors, added randomwalkers

+++ GeneralizedHarmonic.java
@@ -3,9 +3,8 @@ public class GeneralizedHarmonic {
         int n = Integer.parseInt(args[0]);
         int r = Integer.parseInt(args[1]);
         double sum = 0.0;
-
-        for (int i = 1; i <= n; i++) {
-            sum += (1.0 / Math.pow(i, r));
+        for (double i = 1.0; i <= n; i++) {
+            sum += 1.0 / Math.pow(i, r);
         }
         System.out.println(sum);
     }
